+++ src/components/Header/Header.js
@@ -1,13 +1,12 @@
-import { Component } from 'react';
-import TokenService from '../../services/token-service';
+import React, { Component } from 'react';
 
-export default class Header extends Component {
-  render() {
-    return (
-      <nav className='header'>
-        <Link to='/'><h1>Endless Encounters</h1></Link>
-        
-      </nav>
-    )
-  }
+export default function Header() {
+  return (
+    <nav className="border flex flex-row-wrap flex-1 width-full">
+      <div className="flex-1">User Name</div>
+      <div className="flex-1">Home</div>
+      <div className="flex-1">About</div>
+      <div className="flex-1">Login/Logout</div>
+    </nav>
+  )
 }
