+++ src/components/LoginForm/LoginForm.js
@@ -1,38 +1,30 @@
-import { Component } from 'react';
-import TokenService from '../../services/token-service';
+import React from 'react';
 
-export default class LoginForm extends Component {
-  handleSubmit = (e) => {
-    e.preventDefault();
-    const { email, username, password } = e.target;
-
-    console.log(email, username, password);
-    email = '';
-    username = '';
-    password = '';
-
-    //Insert fetch to back end here
-    TokenService.saveAuthToken(TokenService.makeAuthToken(username.value, password.value));
-  }
-
-  render() {
+export default function Login() {
     return (
-      <form className='login-form'
-        onSubmit={this.handleSubmit}>
-          <div className='input-container'>
-            <label htmlFor='email'>Email: </label>
-            <input name='email' type='text' placeholder='Something@bozo.com'/>
-            <label htmlFor='username'>Username: </label>
-            <input name='username' type='text' placeholder='AwesomeKnite420202020'/>
-            <label htmlFor='password'>Password: </label>
-            <input name='password' type='password' placeholder='Asdfsd32'/>
-          </div>
-          <div className='button-container'>
-            <button className='login-button' type='submit'>Login</button>
-            <button className='facebook-button'>Login with Facebook</button>
-            <Link to='/register'>Don't have an account?</Link>
-          </div>
-      </form>
+      <main className="flex flex-col center flex-1 width-full">
+        <div className="flex-1 flex  center p-2 mt-4 justify-stretch">
+
+          <form className="boxshadow rounded-lighter flex-1 flex-grow flex flex-col center  m-2">
+
+            <div className="flex flex-row flex-1 mx-2 max-width-80 font-lg">
+              <label htmlFor="username" className="flex-1">User Name</label>
+              <input type="text" name="username" id='username'
+                className="flex-1 width-full font-lg border-none border-b" />
+            </div>
+
+            <div className="flex flex-row flex-1 mx-2 max-width-80 font-lg">
+              <label htmlFor="userpass" className="flex-1">Password</label>
+              <input type="password" name="userpass" id="userpass"
+                className="flex-1 width-full font-lg border-none border-b" />
+            </div>
+
+            <div className="flex-1 flex flex-col justify-around">
+              <button className="flex-1 m-1 boxshadow-light border-none rounded-lighter">Login</button>
+              <button className="flex-2 m-1 boxshadow-light border-none rounded-lighter">Login With Facebook</button>
+            </div>
+          </form>
+        </div>
+      </main>
     )
-  }
 }
