+++ src/routes/About/About.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class About extends React.Component {
+  render() {
+    return (
+      <main>
+        
+      </main>
+    )
+  }
+}
