+++ src/routes/Dashboard/Dashboard.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Dashboard extends React.Component {
+  render() {
+    return (
+      <main>
+        
+      </main>
+    )
+  }
+}
