+++ src/routes/Home/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Home extends React.Component {
+  render() {
+    return (
+      <main>
+        
+      </main>
+    )
+  }
+}
