+++ src/routes/Register/Register.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Register extends React.Component {
+  render() {
+    return (
+      <main>
+        
+      </main>
+    )
+  }
+}
