+++ src/services/auth-api-service.js
@@ -1,4 +1,3 @@
-import TokenService from './token-service';
 import config from '../config';
 
 const AuthApiService = {
